+++ caffe2/opt/backend_transformer_base.h
@@ -1,0 +1,70 @@
+#pragma once
+
+#include "caffe2/core/common.h"
+#include "caffe2/core/workspace.h"
+#include "caffe2/opt/bound_shape_inferencer.h"
+#include "caffe2/proto/caffe2_pb.h"
+
+#include <string>
+#include <unordered_map>
+#include <vector>
+
+namespace caffe2 {
+namespace {
+constexpr char kNetPos[] = "net_pos";
+constexpr char kModelId[] = "model_id";
+} // namespace
+
+// This class contains some common functions for backend lowering and graph
+// cutting
+class BackendTransformerBase {
+ public:
+  BackendTransformerBase() {}
+  virtual ~BackendTransformerBase() {}
+
+  const std::unordered_map<std::string, std::string>& input_mapping() const {
+    return input_mapping_;
+  }
+
+  const std::unordered_map<std::string, std::string>& reverse_input_mapping()
+      const {
+    return reverse_input_mapping_;
+  }
+
+  virtual void transform(
+      Workspace* ws,
+      NetDef* pred_net,
+      const std::vector<std::string>& weight_names,
+      const std::unordered_map<std::string, TensorShape>& shape_hints,
+      const std::unordered_set<int>& blacklisted_ops) = 0;
+
+ protected:
+  // get model ID from the NetDef
+  std::string getModelId(const NetDef& net);
+
+  // SSA rewrite the net and return name mapping
+  std::unordered_map<std::string, TensorShape> ssaRewriteAndMapNames(
+      Workspace* ws,
+      NetDef* pred_net,
+      const std::unordered_set<std::string>& weights,
+      const std::unordered_map<std::string, TensorShape>& input_shape_hints);
+
+  // Wrap TensorShape into TensorProto
+  TensorProto wrapShapeInfoIntoTensorProto(
+      const std::string& name,
+      const ShapeInfo& shape_info);
+
+  // Do bound shape inference and collect shape infos
+  ShapeInfoMap inferShapes(
+      Workspace* ws,
+      NetDef* pred_net,
+      const std::unordered_map<std::string, TensorShape>& shape_hints_mapped,
+      const BoundShapeSpec& spec);
+
+  // Input mapping of input name -> original input name
+  std::unordered_map<std::string, std::string> input_mapping_;
+
+  // Input mapping of orignal input name -> input name
+  std::unordered_map<std::string, std::string> reverse_input_mapping_;
+};
+} // namespace caffe2
