+++ caffe2/opt/onnxifi_transformer.cc
@@ -16,30 +16,8 @@
 namespace caffe2 {
 
 namespace {
-
-using ShapeInfoMap = std::unordered_map<std::string, ShapeInfo>;
-
-const std::string kNetPos("net_pos");
-const std::string kModelId("model_id");
 const std::string kRealBatchSizeBlob("real_batch_size");
 constexpr size_t kBufferSize = 64;
-
-void AnnotateOpIndex(NetDef* net) {
-  int i = 0;
-  for (auto& op : *(net->mutable_op())) {
-    AddArgument(kNetPos, i++, &op);
-  }
-}
-
-std::string GetModelId(const NetDef& net) {
-  static std::atomic<size_t> seq_id{0};
-  auto model_id =
-      ArgumentHelper(net).GetSingleArgument<std::string>("model_id", "");
-  if (model_id.empty()) {
-    model_id = "unnamed_" + c10::to_string(seq_id++);
-  }
-  return model_id;
-}
 
 // Convert ShapeInfo map to TensorShape map
 std::unordered_map<std::string, TensorShape> StripShapeInfoMap(
@@ -49,19 +27,6 @@ std::unordered_map<std::string, TensorShape> StripShapeInfoMap(
     shape_map.emplace(kv.first, kv.second.shape);
   }
   return shape_map;
-}
-
-// Wrap TensorShape into TensorProto
-TensorProto WrapShapeInfoIntoTensorProto(
-    const std::string& name,
-    const ShapeInfo& shape_info) {
-  TensorProto t;
-  t.set_name(name);
-  t.set_data_type(shape_info.shape.data_type());
-  for (const auto i : shape_info.shape.dims()) {
-    t.add_dims(i);
-  }
-  return t;
 }
 
 uint64_t OnnxifiDataType(caffe2::TensorProto::DataType t) {
@@ -85,38 +50,6 @@ uint64_t OnnxifiDataType(caffe2::TensorProto::DataType t) {
 #undef CAFFE2_TO_ONNXIFI_TYPE
 }
 
-ShapeInfoMap InferShapes(
-    Workspace* ws,
-    NetDef* pred_net,
-    std::unordered_map<std::string, TensorShape>* shape_hints_mapped,
-    const BoundShapeSpec& spec) {
-  ShapeInfoMap shape_map;
-  // Populate shapes from workplace
-  const std::vector<std::string> ws_blobs = ws->Blobs();
-  for (const auto& s : ws_blobs) {
-    auto shape_info = getShapeInfoFromBlob(ws->GetBlob(s));
-    if (shape_info.dim_type != ShapeInfo::DimType::UNKNOWN) {
-      shape_map[s] = shape_info;
-    }
-  }
-  for (const auto& kv : *shape_hints_mapped) {
-    shape_map.emplace(
-        std::piecewise_construct,
-        std::forward_as_tuple(kv.first),
-        std::forward_as_tuple(ShapeInfo::DimType::CONSTANT, kv.second));
-  }
-  BoundShapeInferencer eng(spec);
-  eng.InferBoundShapeAndType(*pred_net, shape_map);
-  const auto& out_map = eng.shape_info();
-
-  for (const auto& kv : out_map) {
-    shape_map.emplace(
-        std::piecewise_construct,
-        std::forward_as_tuple(kv.first),
-        std::forward_as_tuple(kv.second.dim_type, kv.second.shape));
-  }
-  return shape_map;
-}
 
 std::vector<::ONNX_NAMESPACE::ValueInfoProto> ConvertToValueInfo(
     const std::vector<std::string>& names,
@@ -372,7 +305,7 @@ NetDef ComposeResultNet(
 } // namespace
 
 OnnxifiTransformer::OnnxifiTransformer(const OnnxifiTransformerOptions& opts)
-    : opts_(opts) {
+    : BackendTransformerBase(), opts_(opts) {
   lib_ = onnx::initOnnxifiLibrary();
   CAFFE_ENFORCE(lib_, "Cannot initialize ONNXIFI library");
   CAFFE_ENFORCE_EQ(
@@ -516,7 +449,7 @@ NetDef OnnxifiTransformer::SubnetToOnnxifiOpViaC2(
     }
     onnxifi_net.add_external_input(input);
     shape_arg->mutable_tensors()->Add()->CopyFrom(
-        WrapShapeInfoIntoTensorProto(input, shape_hints.at(i)));
+        wrapShapeInfoIntoTensorProto(input, shape_hints.at(i)));
   }
 
   // Compute output shape hints
@@ -703,58 +636,6 @@ NetDef OnnxifiTransformer::SubnetToOnnxifiOpViaOnnx(
   return net_opt;
 }
 
-std::unordered_map<std::string, TensorShape>
-OnnxifiTransformer::SsaRewriteAndMapNames(
-    Workspace* ws,
-    NetDef* pred_net,
-    const std::unordered_set<std::string>& weights,
-    const std::unordered_map<std::string, TensorShape>& input_shape_hints) {
-  // Make sure weights do not contain output of any op.
-  for (const auto& op : pred_net->op()) {
-    for (const auto& output : op.output()) {
-      CAFFE_ENFORCE_EQ(
-          weights.count(output),
-          0,
-          "Weight ",
-          output,
-          " shouldn't appear in the output");
-    }
-  }
-  input_mapping_ = onnx::SsaRewrite(nullptr, pred_net, weights);
-  // Annote the ops with net position
-  AnnotateOpIndex(pred_net);
-
-  // Need to add mapping for weights. This will be used to create new workspace
-  // with mapped weights.
-  for (const auto& w : weights) {
-    input_mapping_.emplace(w, w);
-  }
-
-  // Since we are going to create a mapped workspace, we need to make sure that
-  // the parent workspace has the mapped blob names. If the blobs don't exist
-  // (usually such blobs are input tensor names), we exclude them from mapping.
-  std::vector<std::string> exclude_mapping;
-  for (const auto kv : input_mapping_) {
-    reverse_input_mapping_.emplace(kv.second, kv.first);
-    if (!ws->HasBlob(kv.second)) {
-      exclude_mapping.emplace_back(kv.first);
-    }
-  }
-  for (const auto& i : exclude_mapping) {
-    input_mapping_.erase(i);
-  }
-  std::unordered_map<std::string, TensorShape> shape_hints_mapped;
-  for (const auto& kv : input_shape_hints) {
-    const auto it = reverse_input_mapping_.find(kv.first);
-    if (it != reverse_input_mapping_.end()) {
-      shape_hints_mapped.emplace(it->second, kv.second);
-    } else {
-      shape_hints_mapped.emplace(kv.first, kv.second);
-    }
-  }
-  return shape_hints_mapped;
-}
-
 NetDef OnnxifiTransformer::TransformViaC2(
     NetDef* pred_net,
     const std::unordered_set<std::string>& weights,
@@ -779,7 +660,7 @@ NetDef OnnxifiTransformer::TransformViaC2(
   }
   onnxBackendID backend_id = backend_ids_[idx_];
 
-  auto c2_supports = [&shape_hints, &blacklisted_ops, backend, backend_id](
+  auto c2_supports = [this, &shape_hints, &blacklisted_ops, backend, backend_id](
                          const caffe2::OperatorDef& op) {
     try {
       int pos =
@@ -812,7 +693,7 @@ NetDef OnnxifiTransformer::TransformViaC2(
           return false;
         }
         shape_arg->mutable_tensors()->Add()->CopyFrom(
-            WrapShapeInfoIntoTensorProto(i, it->second));
+            wrapShapeInfoIntoTensorProto(i, it->second));
       }
       shape_arg = net.add_arg();
       shape_arg->set_name("output_shape_info");
@@ -822,7 +703,7 @@ NetDef OnnxifiTransformer::TransformViaC2(
           return false;
         }
         shape_arg->mutable_tensors()->Add()->CopyFrom(
-            WrapShapeInfoIntoTensorProto(i, it->second));
+            wrapShapeInfoIntoTensorProto(i, it->second));
       }
 
       std::string c2_model_str;
@@ -978,7 +859,7 @@ NetDef OnnxifiTransformer::TransformViaOnnx(
 
 // Cutting off the runnable part and replace with ONNXIFI ops. Asssume the nets
 // were topologically sorted
-void OnnxifiTransformer::Transform(
+void OnnxifiTransformer::transform(
     Workspace* ws,
     NetDef* pred_net,
     const std::vector<std::string>& weight_names,
@@ -987,8 +868,8 @@ void OnnxifiTransformer::Transform(
   CAFFE_ENFORCE(ws);
   CAFFE_ENFORCE(pred_net, "Predict net cannot be nullptr");
 
-  // Get model id and  reset Onnxifi op id to 0
-  model_id_ = GetModelId(*pred_net);
+  // Get model id and reset Onnxifi op id to 0
+  model_id_ = getModelId(*pred_net);
   onnxifi_op_id_ = 0;
 
   std::unordered_set<std::string> weights(
@@ -996,12 +877,12 @@ void OnnxifiTransformer::Transform(
 
   // SSA Rewrite the net
   auto shape_hints_mapped =
-      SsaRewriteAndMapNames(ws, pred_net, weights, input_shape_hints);
+      ssaRewriteAndMapNames(ws, pred_net, weights, input_shape_hints);
 
   // Populate shape info
   Workspace mapped_ws(ws, input_mapping_);
-  ShapeInfoMap shape_hints = InferShapes(
-      &mapped_ws, pred_net, &shape_hints_mapped, opts_.bound_shape_spec);
+  ShapeInfoMap shape_hints = inferShapes(
+      &mapped_ws, pred_net, shape_hints_mapped, opts_.bound_shape_spec);
 
   // Transform the net
   NetDef net_opt = opts_.use_onnx
