+++ caffe2/opt/onnxifi_transformer.h
@@ -7,12 +7,9 @@
 
 #include "onnx/onnx_pb.h"
 
-#include "caffe2/core/common.h"
 #include "caffe2/core/operator.h"
-#include "caffe2/core/workspace.h"
 #include "caffe2/onnx/onnxifi_init.h"
-#include "caffe2/opt/bound_shape_inferencer.h"
-#include "caffe2/proto/caffe2_pb.h"
+#include "caffe2/opt/backend_transformer_base.h"
 
 namespace caffe2 {
 namespace onnx {
@@ -31,26 +28,17 @@ struct OnnxifiTransformerOptions {
   BoundShapeSpec bound_shape_spec;
 };
 
-class CAFFE2_API OnnxifiTransformer final {
+class CAFFE2_API OnnxifiTransformer final : public BackendTransformerBase {
  public:
   explicit OnnxifiTransformer(const OnnxifiTransformerOptions& opts);
-  ~OnnxifiTransformer();
+  ~OnnxifiTransformer() override;
 
-  void Transform(
+  void transform(
       Workspace* ws,
       NetDef* pred_net,
       const std::vector<std::string>& weight_names,
       const std::unordered_map<std::string, TensorShape>& shape_hints,
-      const std::unordered_set<int>& blacklisted_ops);
-
-  const std::unordered_map<std::string, std::string>& input_mapping() const {
-    return input_mapping_;
-  }
-
-  const std::unordered_map<std::string, std::string>& reverse_input_mapping()
-      const {
-    return reverse_input_mapping_;
-  }
+      const std::unordered_set<int>& blacklisted_ops) override;
 
  private:
   // Since we create new tensors during the conversion process, we actually need
@@ -79,12 +67,6 @@ class CAFFE2_API OnnxifiTransformer final {
       const std::unordered_set<std::string>& initialization_list,
       const std::vector<std::string>& external_inputs,
       const std::vector<std::string>& external_outputs);
-
-  std::unordered_map<std::string, TensorShape> SsaRewriteAndMapNames(
-      Workspace* ws,
-      NetDef* pred_net,
-      const std::unordered_set<std::string>& weights,
-      const std::unordered_map<std::string, TensorShape>& input_shape_hints);
 
   // Transform by passing C2 proto to backend
   NetDef TransformViaC2(
@@ -121,11 +103,5 @@ class CAFFE2_API OnnxifiTransformer final {
 
   // Backned IDs
   std::vector<onnxBackendID> backend_ids_;
-
-  // Input mapping of input name -> original input name
-  std::unordered_map<std::string, std::string> input_mapping_;
-
-  // Input mapping of orignal input name -> input name
-  std::unordered_map<std::string, std::string> reverse_input_mapping_;
 };
 } // namespace caffe2
